requirements_dev.txt
wheel>=0.41
watchdog==3.0.0
flake8==6.1.0
tox==4.11.3
coverage==7.3.2
sphinx==7.2.6
twine==4.0.2
pytest==7.4.3
black==23.11.0
pyqir==0.10.0
qiskit-terra>=0.19.2,<1.0